requirements.txt
alembic==1.2.1
Click==7.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.0
marshmallow-sqlalchemy==0.19.0
mysqlclient==1.4.4
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.16.0
